Refactor, add javadocs
